# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyPets",
        "description": "bark",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "auth/password/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send recovery password link",
                "description": "Send link for recovery password.",
                "operationId": "forgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link sended",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User access errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User is registered through Facebook.|You are not pet owner."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Invalid data"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recovery password",
                "description": "Recovery password",
                "operationId": "resetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Reset password token",
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Access token, pet owner and pet info.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "owner": {
                                                    "properties": {
                                                        "first_name": {
                                                            "type": "string"
                                                        },
                                                        "last_name": {
                                                            "type": "string"
                                                        },
                                                        "gender": {
                                                            "type": "string"
                                                        },
                                                        "age": {
                                                            "type": "integer"
                                                        },
                                                        "birthday": {
                                                            "type": "date"
                                                        },
                                                        "occupation": {
                                                            "type": "string"
                                                        },
                                                        "hobbies": {
                                                            "type": "string"
                                                        },
                                                        "pets_owned": {
                                                            "type": "string"
                                                        },
                                                        "profile_picture": {
                                                            "type": "string"
                                                        },
                                                        "favorite_park": {
                                                            "type": "string"
                                                        },
                                                        "pet": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "gender": {
                                                                    "type": "string"
                                                                },
                                                                "size": {
                                                                    "type": "string"
                                                                },
                                                                "primary_breed": {
                                                                    "type": "string"
                                                                },
                                                                "secondary_breed": {
                                                                    "type": "string"
                                                                },
                                                                "age": {
                                                                    "type": "integer"
                                                                },
                                                                "profile_picture": {
                                                                    "type": "string"
                                                                },
                                                                "friendliness": {
                                                                    "type": "integer"
                                                                },
                                                                "activity_level": {
                                                                    "type": "integer"
                                                                },
                                                                "noise_level": {
                                                                    "type": "integer"
                                                                },
                                                                "odebience_level": {
                                                                    "type": "integer"
                                                                },
                                                                "fetchability": {
                                                                    "type": "integer"
                                                                },
                                                                "swimability": {
                                                                    "type": "integer"
                                                                },
                                                                "city": {
                                                                    "type": "string"
                                                                },
                                                                "state": {
                                                                    "type": "string"
                                                                },
                                                                "like": {
                                                                    "type": "string"
                                                                },
                                                                "dislike": {
                                                                    "type": "string"
                                                                },
                                                                "favorite_toys": {
                                                                    "type": "string"
                                                                },
                                                                "fears": {
                                                                    "type": "string"
                                                                },
                                                                "favorite_places": {
                                                                    "type": "string"
                                                                },
                                                                "spayed": {
                                                                    "type": "string"
                                                                },
                                                                "birthday": {
                                                                    "type": "date"
                                                                },
                                                                "pictures": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "id": {
                                                                                "type": "integer"
                                                                            },
                                                                            "picture": {
                                                                                "type": "string"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User access error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not pet owner."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Token parse error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid user.|Invalid password.|Invalid token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Invalid data"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}